# docs/swagger.yaml
basePath: /
definitions:
  controllers.approveBudgetRequest:
    properties:
      approved_by_name:
        type: string
    type: object
  controllers.authResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      role:
        type: string
      token_type:
        type: string
      user_id:
        type: string
    type: object
...
  controllers.clientServiceOrderResponse:
    properties:
      budget_status:
        type: string
      budget_total_cents:
        type: integer
      code:
        type: string
      opened_at:
        type: string
      status:
        type: string
    type: object
...
  controllers.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  controllers.rejectBudgetRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
  dto.CreateServiceRequest:
    properties:
      active:
        type: boolean
      base_price_cents:
        type: integer
      description:
        type: string
      estimated_minutes:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  dto.CreateVehicleRequest:
    properties:
      brand:
        type: string
      chassis:
        type: string
      color:
        type: string
      manufacture_year:
        type: integer
      mileage:
        type: integer
      model:
        type: string
      model_year:
        type: integer
      notes:
        type: string
      plate:
        type: string
    required:
    - brand
    - model
    - model_year
    - plate
    type: object
...
  dto.ServiceResponse:
    properties:
      active:
        type: boolean
      base_price_cents:
        type: integer
      description:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
  dto.VehicleResponse:
    properties:
      brand:
        type: string
      chassis:
        type: string
      client_id:
        type: string
      color:
        type: string
      id:
        type: string
      manufacture_year:
        type: integer
      mileage:
        type: integer
      model:
        type: string
      model_year:
        type: integer
      notes:
        type: string
      plate:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Tech Challenge project.
  title: Tech Challenge API
  version: "1.0"
paths:
...
  /admin/clients/{id}/vehicles:
    get:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.VehicleResponse'
            type: array
      summary: List vehicles by client
      tags:
      - admin-vehicles
    post:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Vehicle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVehicleRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.VehicleResponse'
      summary: Create vehicle for client
      tags:
      - admin-vehicles
...
  /admin/services:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ServiceResponse'
            type: array
      summary: List services
      tags:
      - admin-services
    post:
      parameters:
      - description: Service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateServiceRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ServiceResponse'
      summary: Create service
      tags:
      - admin-services
  /admin/services/{id}:
    delete:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete service
      tags:
      - admin-services
    get:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ServiceResponse'
      summary: Get service
      tags:
      - admin-services
    put:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateServiceRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ServiceResponse'
      summary: Update service
      tags:
      - admin-services
  /admin/vehicles/{id}:
    delete:
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete vehicle
      tags:
      - admin-vehicles
    get:
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VehicleResponse'
      summary: Get vehicle
      tags:
      - admin-vehicles
    put:
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      - description: Vehicle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVehicleRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VehicleResponse'
      summary: Update vehicle
      tags:
      - admin-vehicles
...
  /auth/register:
    post:
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.registerRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.authResponse'
      summary: Register a new user
      tags:
      - auth
  /client/service-orders/{code}:
    get:
      parameters:
      - description: CPF/CNPJ
        in: query
        name: document_number
        required: true
        type: string
      - description: Service Order Code
        in: path
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.clientServiceOrderResponse'
      summary: Get service order progress (client)
      tags:
      - client-service-orders
  /client/service-orders/{code}/budget/approve:
    post:
      parameters:
      - description: CPF/CNPJ
        in: query
        name: document_number
        required: true
        type: string
      - description: Service Order Code
        in: path
        name: code
        required: true
        type: string
      - description: Approval info
        in: body
        name: request
        schema:
          $ref: '#/definitions/controllers.approveBudgetRequest'
      responses:
        "204":
          description: No Content
      summary: Approve latest budget
      tags:
      - client-service-orders
  /client/service-orders/{code}/budget/reject:
    post:
      parameters:
      - description: CPF/CNPJ
        in: query
        name: document_number
        required: true
        type: string
      - description: Service Order Code
        in: path
        name: code
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.rejectBudgetRequest'
      responses:
        "204":
          description: No Content
      summary: Reject latest budget
      tags:
      - client-service-orders
...
securityDefinitions:
  BearerAuth:
    description: Type `Bearer ` followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"